pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invest-researcher"
version = "0.1.0"
description = "A-share value investing 5-step analysis web app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openpyxl",
    "fastapi",
    "uvicorn[standard]",
    "akshare",
    "openai",
    "anthropic",
    "pydantic-settings",
    "sse-starlette",
    "httpx",
    "diskcache",
]

[project.scripts]
start = "scripts.dev:main"

[tool.hatch.build.targets.wheel]
packages = ["backend", "scripts"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
]
